recycle/requirements.txt
pyyaml>=6.0 